+++ keras_cv/models/__init__.py
@@ -30,28 +30,28 @@ from keras_cv.models.backbones.csp_darknet.csp_darknet_backbone import (
 from keras_cv.models.backbones.csp_darknet.csp_darknet_backbone import (
     CSPDarkNetXLBackbone,
 )
-from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_aliases import (
     EfficientNetV2B0Backbone,
 )
-from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_aliases import (
     EfficientNetV2B1Backbone,
 )
-from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_aliases import (
     EfficientNetV2B2Backbone,
 )
-from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_aliases import (
     EfficientNetV2B3Backbone,
 )
-from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_aliases import (
     EfficientNetV2Backbone,
 )
-from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_aliases import (
     EfficientNetV2LBackbone,
 )
-from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_aliases import (
     EfficientNetV2MBackbone,
 )
-from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_aliases import (
     EfficientNetV2SBackbone,
 )
 from keras_cv.models.backbones.mobilenet_v3.mobilenet_v3_backbone import (
