+++ keras_cv/models/backbones/efficientnet_v2/efficientnet_v2_backbone.py
@@ -297,24 +297,6 @@ class EfficientNetV2Backbone(Backbone):
         return copy.deepcopy(backbone_presets_with_weights)
 
 
-BASE_DOCSTRING = """Instantiates the {name} architecture.
-
-    Reference:
-    - [EfficientNetV2: Smaller Models and Faster Training](https://arxiv.org/abs/2104.00298)
-      (ICML 2021)
-
-    Args:
-        include_rescaling: bool, whether to rescale the inputs. If set
-            to `True`, inputs will be passed through a `Rescaling(1/255.0)`
-            layer.
-        input_shape: optional shape tuple, defaults to (None, None, 3).
-        input_tensor: optional Keras tensor (i.e. output of `layers.Input()`)
-            to use as image input for the model.
-    Returns:
-      A `keras.Model` instance.
-"""  # noqa: E501
-
-
 def conv_kernel_initializer(scale=2.0):
     return keras.initializers.VarianceScaling(
         scale=scale, mode="fan_out", distribution="truncated_normal"
@@ -374,270 +356,3 @@ def get_block_conv(
         survival_probability=survival_probability,
         name=name,
     )
-
-
-class EfficientNetV2SBackbone(EfficientNetV2Backbone):
-    def __new__(
-        cls,
-        include_rescaling=True,
-        input_shape=(None, None, 3),
-        input_tensor=None,
-        **kwargs,
-    ):
-        # Pack args in kwargs
-        kwargs.update(
-            {
-                "include_rescaling": include_rescaling,
-                "input_shape": input_shape,
-                "input_tensor": input_tensor,
-            }
-        )
-        return EfficientNetV2Backbone.from_preset("efficientnetv2-s", **kwargs)
-
-    @classproperty
-    def presets(cls):
-        """Dictionary of preset names and configurations."""
-        return {
-            "efficientnetv2-s": copy.deepcopy(
-                backbone_presets["efficientnetv2-s"]
-            ),
-        }
-
-    @classproperty
-    def presets_with_weights(cls):
-        """Dictionary of preset names and configurations that include
-        weights."""
-        return {
-            "efficientnetv2-s_imagenet": copy.deepcopy(
-                backbone_presets["efficientnetv2-s_imagenet"]
-            ),
-        }
-
-
-class EfficientNetV2MBackbone(EfficientNetV2Backbone):
-    def __new__(
-        cls,
-        include_rescaling=True,
-        input_shape=(None, None, 3),
-        input_tensor=None,
-        **kwargs,
-    ):
-        # Pack args in kwargs
-        kwargs.update(
-            {
-                "include_rescaling": include_rescaling,
-                "input_shape": input_shape,
-                "input_tensor": input_tensor,
-            }
-        )
-        return EfficientNetV2Backbone.from_preset("efficientnetv2-m", **kwargs)
-
-    @classproperty
-    def presets(cls):
-        """Dictionary of preset names and configurations."""
-        return {
-            "efficientnetv2-m": copy.deepcopy(
-                backbone_presets["efficientnetv2-m"]
-            ),
-        }
-
-    @classproperty
-    def presets_with_weights(cls):
-        """Dictionary of preset names and configurations that include
-        weights."""
-        return {}
-
-
-class EfficientNetV2LBackbone(EfficientNetV2Backbone):
-    def __new__(
-        cls,
-        include_rescaling=True,
-        input_shape=(None, None, 3),
-        input_tensor=None,
-        **kwargs,
-    ):
-        # Pack args in kwargs
-        kwargs.update(
-            {
-                "include_rescaling": include_rescaling,
-                "input_shape": input_shape,
-                "input_tensor": input_tensor,
-            }
-        )
-        return EfficientNetV2Backbone.from_preset("efficientnetv2-l", **kwargs)
-
-    @classproperty
-    def presets(cls):
-        """Dictionary of preset names and configurations."""
-        return {
-            "efficientnetv2-l": copy.deepcopy(
-                backbone_presets["efficientnetv2-l"]
-            ),
-        }
-
-    @classproperty
-    def presets_with_weights(cls):
-        """Dictionary of preset names and configurations that include
-        weights."""
-        return {}
-
-
-class EfficientNetV2B0Backbone(EfficientNetV2Backbone):
-    def __new__(
-        cls,
-        include_rescaling=True,
-        input_shape=(None, None, 3),
-        input_tensor=None,
-        **kwargs,
-    ):
-        # Pack args in kwargs
-        kwargs.update(
-            {
-                "include_rescaling": include_rescaling,
-                "input_shape": input_shape,
-                "input_tensor": input_tensor,
-            }
-        )
-        return EfficientNetV2Backbone.from_preset("efficientnetv2-b0", **kwargs)
-
-    @classproperty
-    def presets(cls):
-        """Dictionary of preset names and configurations."""
-        return {
-            "efficientnetv2-b0": copy.deepcopy(
-                backbone_presets["efficientnetv2-b0"]
-            ),
-        }
-
-    @classproperty
-    def presets_with_weights(cls):
-        """Dictionary of preset names and configurations that include
-        weights."""
-        return {
-            "efficientnetv2-b0_imagenet": copy.deepcopy(
-                backbone_presets["efficientnetv2-b0_imagenet"]
-            ),
-        }
-
-
-class EfficientNetV2B1Backbone(EfficientNetV2Backbone):
-    def __new__(
-        cls,
-        include_rescaling=True,
-        input_shape=(None, None, 3),
-        input_tensor=None,
-        **kwargs,
-    ):
-        # Pack args in kwargs
-        kwargs.update(
-            {
-                "include_rescaling": include_rescaling,
-                "input_shape": input_shape,
-                "input_tensor": input_tensor,
-            }
-        )
-        return EfficientNetV2Backbone.from_preset("efficientnetv2-b1", **kwargs)
-
-    @classproperty
-    def presets(cls):
-        """Dictionary of preset names and configurations."""
-        return {
-            "efficientnetv2-b1": copy.deepcopy(
-                backbone_presets["efficientnetv2-b1"]
-            ),
-        }
-
-    @classproperty
-    def presets_with_weights(cls):
-        """Dictionary of preset names and configurations that include
-        weights."""
-        return {
-            "efficientnetv2-b1_imagenet": copy.deepcopy(
-                backbone_presets["efficientnetv2-b1_imagenet"]
-            ),
-        }
-
-
-class EfficientNetV2B2Backbone(EfficientNetV2Backbone):
-    def __new__(
-        cls,
-        include_rescaling=True,
-        input_shape=(None, None, 3),
-        input_tensor=None,
-        **kwargs,
-    ):
-        # Pack args in kwargs
-        kwargs.update(
-            {
-                "include_rescaling": include_rescaling,
-                "input_shape": input_shape,
-                "input_tensor": input_tensor,
-            }
-        )
-        return EfficientNetV2Backbone.from_preset("efficientnetv2-b2", **kwargs)
-
-    @classproperty
-    def presets(cls):
-        """Dictionary of preset names and configurations."""
-        return {
-            "efficientnetv2-b2": copy.deepcopy(
-                backbone_presets["efficientnetv2-b2"]
-            ),
-        }
-
-    @classproperty
-    def presets_with_weights(cls):
-        """Dictionary of preset names and configurations that include
-        weights."""
-        return {
-            "efficientnetv2-b2_imagenet": copy.deepcopy(
-                backbone_presets["efficientnetv2-b2_imagenet"]
-            ),
-        }
-
-
-class EfficientNetV2B3Backbone(EfficientNetV2Backbone):
-    def __new__(
-        cls,
-        include_rescaling=True,
-        input_shape=(None, None, 3),
-        input_tensor=None,
-        **kwargs,
-    ):
-        # Pack args in kwargs
-        kwargs.update(
-            {
-                "include_rescaling": include_rescaling,
-                "input_shape": input_shape,
-                "input_tensor": input_tensor,
-            }
-        )
-        return EfficientNetV2Backbone.from_preset("efficientnetv2-b3", **kwargs)
-
-    @classproperty
-    def presets(cls):
-        """Dictionary of preset names and configurations."""
-        return {}
-
-    @classproperty
-    def presets_with_weights(cls):
-        """Dictionary of preset names and configurations that include
-        weights."""
-        return {}
-
-
-EfficientNetV2B0Backbone.__doc__ = BASE_DOCSTRING.format(
-    name="EfficientNetV2B0"
-)
-EfficientNetV2B1Backbone.__doc__ = BASE_DOCSTRING.format(
-    name="EfficientNetV2B1"
-)
-EfficientNetV2B2Backbone.__doc__ = BASE_DOCSTRING.format(
-    name="EfficientNetV2B2"
-)
-EfficientNetV2B3Backbone.__doc__ = BASE_DOCSTRING.format(
-    name="EfficientNetV2B3"
-)
-EfficientNetV2SBackbone.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2S")
-EfficientNetV2MBackbone.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2M")
-EfficientNetV2LBackbone.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2L")
