+++ keras_cv/models/backbones/efficientnet_v2/efficientnet_v2_backbone_test.py
@@ -18,11 +18,11 @@ import tensorflow as tf
 from absl.testing import parameterized
 from tensorflow import keras
 
-from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
-    EfficientNetV2Backbone,
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_aliases import (
+    EfficientNetV2SBackbone,
 )
 from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import (
-    EfficientNetV2SBackbone,
+    EfficientNetV2Backbone,
 )
 from keras_cv.utils.train import get_feature_extractor
